Update test helpers for dropCollections

+++ test/helpers.js
@@ -5,7 +5,7 @@ import { MongoSingleton } from '../src/config/mongodb.config.js'
 
 export const mongoConection = MongoSingleton.getInstance()
 
-export const dropCollections = async () => {
+export const dropAllCollections = async (collectionNames) => {
   const collections = Object.keys(mongoose.connection.collections)
   for (const collectionName of collections) {
     const collection = mongoose.connection.collections[collectionName]
@@ -13,7 +13,10 @@ export const dropCollections = async () => {
   }
 }
 
-export const dropCollection = async (collectionName) => {
-  const collection = mongoose.connection.collections[collectionName]
-  await collection.deleteMany({})
+export const dropCollection = async (...collectionNames) => {
+  for (const collectionName of collectionNames) {
+    const collection = mongoose.connection.collections[collectionName]
+    if (!collection) return console.log(`"${collectionName}" collection not found`)
+    await collection.deleteMany({})
+  }
 }
